Moved out testing codes and fixed permission of some scripts
