+++ liota/utilities/utility.py
